Extracted the SearchBar component logic into a new useSearchBar custom hook
